refactor(argument-resolver): add new logic to ensure token extraction fails
- if `"Bearer "` is missing, throw
- define constants for magic strings
- move a package similar to `webconfig` for better cohesion.
