+++ src/main/kotlin/ecommerce/config/WebConfig.kt
@@ -1,6 +1,5 @@
 package ecommerce.config
 
-import ecommerce.service.LoginMemberArgumentResolver
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.method.support.HandlerMethodArgumentResolver
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
