+++ src/main/kotlin/ecommerce/sql/AuthConstsSQL.kt
@@ -1,0 +1,7 @@
+package ecommerce.sql
+
+object AuthConstsSQL {
+    val HEADER_AUTHORIZATION = "Authorization"
+    val HEADER_PREFIX = "Bearer "
+
+}
